Reuse existing ServerInfo object when reloading config
Prior to this command, greload would replace the ServerInfo stored in
bungee, this leads to the /glist counters being out of sync, but also
potentially highlights futher complications in replacing live server
objects, in that code doesn't generally expect/account for it to occur.

This behavior is not as ideal as being able to use the new server info,
which will have updated settings such as the motd, however, this change
offers better expected behavior.

We may wish to revisit this in the future, e.g. system property to force
using the new objects at the risk of breaking the player count on
reload?